+++ src/@cdk/components/componente-digital/cdk-componente-digital-card-list/cdk-componente-digital-card-list.component.ts
@@ -9,9 +9,9 @@ import {
 } from '@angular/core';
 
 import {cdkAnimations} from '@cdk/animations';
-import {ComponenteDigital, Documento, DocumentoAvulso, Processo, Tarefa} from '@cdk/models';
-import {classToPlain} from 'class-transformer';
-import {HttpClient, HttpErrorResponse, HttpEventType, HttpRequest} from '@angular/common/http';
+import {ComponenteDigital, Documento, DocumentoAvulso, Processo, Tarefa, Visibilidade} from '@cdk/models';
+import {classToPlain, plainToClass} from 'class-transformer';
+import {HttpClient, HttpErrorResponse, HttpEventType, HttpParams, HttpRequest} from '@angular/common/http';
 import {catchError, last, map, tap} from 'rxjs/operators';
 import {of, Subscription} from 'rxjs';
 import {environment} from 'environments/environment';
@@ -277,10 +277,16 @@ export class CdkComponenteDigitalCardListComponent {
                 componenteDigital.conteudo = conteudo;
                 this._changeDetectorRef.markForCheck();
 
-                const params = classToPlain(componenteDigital);
+                const body = classToPlain(componenteDigital);
+                const params = {};
+                params['populate'] = JSON.stringify([
+                    'documento',
+                    'documento.tipoDocumento'
+                ]);
 
-                const req = new HttpRequest('POST', this.target, params, {
-                    reportProgress: true
+                const req = new HttpRequest('POST', this.target, body, {
+                    reportProgress: true,
+                    params: new HttpParams({fromObject: params})
                 });
 
                 componenteDigital.inProgress = true;
@@ -326,6 +332,8 @@ export class CdkComponenteDigitalCardListComponent {
                             componenteDigital.id = event.body.id;
                             componenteDigital.complete = true;
                             componenteDigital.inProgress = false;
+                            componenteDigital.documento = plainToClass(Documento, event.body.documento);
+                            delete componenteDigital.documento.tarefaOrigem;
                             this.currentFile = null;
                             this._changeDetectorRef.markForCheck();
                             setTimeout(() => {