+++ src/app/main/apps/tarefas/atividade-create-bloco/store/effects/documentos.effects.ts
@@ -13,7 +13,6 @@ import {
     documento as documentoSchema
 } from '@cdk/normalizr';
 import {ActivatedRoute, Router} from '@angular/router';
-import {environment} from 'environments/environment';
 import * as AtividadeBlocoCreateDocumentosActionsAll from '../actions/documentos.actions';
 import {getSelectedTarefas} from '../../../store';
 import * as fromStore from '../index';
@@ -31,17 +30,16 @@ export class AtividadeCreateBlocoDocumentosEffect {
      */
     getDocumentos: Observable<any> = createEffect(() => this._actions.pipe(
         ofType<AtividadeBlocoCreateDocumentosActionsAll.GetDocumentos>(AtividadeBlocoCreateDocumentosActionsAll.GET_DOCUMENTOS_BLOCO),
-        switchMap((action) => {
-
-            const tarefaIds = `in:${action.payload}`;
+        mergeMap((action) => {
+            const tarefaId = `eq:${action.payload}`;
 
             const params = {
                 filter: {
-                    'tarefaOrigem.id': tarefaIds,
+                    'tarefaOrigem.id': tarefaId,
                     'documentoAvulsoRemessa.id': 'isNull',
                     'juntadaAtual': 'isNull'
                 },
-                limit: 10,
+                limit: 25,
                 offset: 0,
                 sort: {
                     criadoEm: 'DESC'
@@ -62,19 +60,23 @@ export class AtividadeCreateBlocoDocumentosEffect {
                 params.limit,
                 params.offset,
                 JSON.stringify(params.sort),
-                JSON.stringify(params.populate));
+                JSON.stringify(params.populate)
+            ).pipe(
+                mergeMap(response => [
+                    new AddData<Documento>({data: response['entities'], schema: documentoSchema}),
+                    new AtividadeBlocoCreateDocumentosActionsAll.GetDocumentosSuccess({
+                        loaded: {
+                            id: action.payload
+                        },
+                        entitiesId: response['entities'].map(documento => documento.id),
+                    })
+                ]),
+                catchError((err) => {
+                    console.log(err);
+                    return of(new AtividadeBlocoCreateDocumentosActionsAll.GetDocumentosFailed(err));
+                })
+            );
         }),
-        mergeMap(response => [
-            new AddData<Documento>({data: response['entities'], schema: documentoSchema}),
-            new AtividadeBlocoCreateDocumentosActionsAll.GetDocumentosSuccess({
-                loaded: true,
-                entitiesId: response['entities'].map(documento => documento.id),
-            })
-        ]),
-        catchError((err) => {
-            console.log(err);
-            return of(new AtividadeBlocoCreateDocumentosActionsAll.GetDocumentosFailed(err));
-        })
     ));
     /**
      * Delete Documento