+++ src/app/main/apps/tarefas/tarefa-detail/atividades/atividade-create/store/reducers/atividade-create.reducer.ts
@@ -6,12 +6,15 @@ export interface AtividadeCreateState {
     errors: any;
 }
 
-export const AtividadeCreateInitialState: AtividadeCreateState = {
+export const atividadeCreateInitialState: AtividadeCreateState = {
     saving: false,
     errors: false
 };
 
-export function AtividadeCreateReducer(state = AtividadeCreateInitialState, action: AtividadeCreateActions.AtividadeCreateActionsAll): AtividadeCreateState {
+export const atividadeCreateReducer = (
+    state = atividadeCreateInitialState,
+    action: AtividadeCreateActions.AtividadeCreateActionsAll
+): AtividadeCreateState => {
     switch (action.type) {
 
         case AtividadeCreateActions.CREATE_ATIVIDADE: {
@@ -47,4 +50,4 @@ export function AtividadeCreateReducer(state = AtividadeCreateInitialState, acti
         default:
             return state;
     }
-}
+};